mastodon: pass dependencies as environment variables
This makes it much easier to override components of the Mastodon
derivation to e.g. use a fork. With this change, one can use something
like the following to install a fork (with the appropriate files):

    let
      src = callPackage ./source.nix { };
    in
      (mastodon.override {
        pname = "hometown";
        version = import ./version.nix;
        srcOverride = src;
        dependenciesDir = ./.;
      }).overrideAttrs (oldAttrs: rec {
        yarnOfflineCache = fetchYarnDeps {
          yarnLock = "${src}/yarn.lock";
          sha256 = "sha256-AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=";
        };
        mastodonModules = oldAttrs.mastodonModules.overrideAttrs (oldModuleAttrs: {
          inherit yarnOfflineCache;
        });
      });

A spiritual successor to f949de4fbb5c5adb7f3ec085206f48a2efdd1d9a