kdePackages.extra-cmake-modules: propagate qdoc by default
A lot of things hard-require the ECM bits that use it now.

Due to some cursed interaction between Qt's use of cmake and our
strictDeps patch, this makes a bunch of builds fall over until
things are moved (or, in some cases, copied) to nativeBuildInputs.

Disabling strictDeps makes things build, but starts leaking build-deps
into the final outputs via the ecm-hook wrapper, so just ADD ALL THE THINGS
for now, and hopefully we can figure out how to do this properly later.