pypy3Packaghes.eventlet: relax greenlet
greenlet is built into pypy but the version number is older.
Since the tests do not run, we cannot be super sure that this works 100%
but some following packages built. Maybe we should bring back greenlet
to pypy as a package but that is a bigger task for another PR.

(cherry picked from commit 8fc7881896ebeaaf86a55783fc5b1e62608b8ced)